requirements.txt
numpy
scikit-allel
matplotlib
pandas
shapely
geopandas
adjustText
neo4j
scipy